Refactor MessageableChannelViewController for better organization and maintainability
- Split MessageableChannelViewController into focused extensions
- Extract NSFW overlay into dedicated view component
- Add skeleton loading view for message list
- Extract notification banner logic
- Improve code organization and readability
