+++ .gitignore
@@ -6,3 +6,4 @@ DerivedData/
 .build/
 Packages/
 Pods/
+CrashLog.txt
