+++ Revolt/Pages/Channel/Messagable/MessageableChannelViewController.swift
@@ -69,8 +69,8 @@ class MessageableChannelViewController: UIViewController, UITextFieldDelegate,
     var lastManualScrollTime: Date?
     var lastManualScrollUpTime: Date?
     var scrollProtectionTimer: Timer?
-    private var scrollCheckTimer: Timer?
-    private var contentSizeObserverRegistered = false
+    var scrollCheckTimer: Timer?
+    var contentSizeObserverRegistered = false
     var lastScrollToBottomTime: Date?
     let scrollDebounceInterval: TimeInterval = 2.0
     var networkErrorCooldown: TimeInterval = 5.0
@@ -106,7 +106,7 @@ class MessageableChannelViewController: UIViewController, UITextFieldDelegate,
     var toggleSidebar: (() -> Void)?
 
     // Track whether we're returning from search to prevent unnecessary cleanup
-    private var isReturningFromSearch: Bool = false
+    var isReturningFromSearch: Bool = false
 
     // Target message ID to scroll to
     var targetMessageId: String? {
@@ -1353,7 +1353,7 @@ class MessageableChannelViewController: UIViewController, UITextFieldDelegate,
     }
 
     // Helper method to log memory usage
-    private func logMemoryUsage(prefix: String) {
+    func logMemoryUsage(prefix: String) {
         var info = mach_task_basic_info()
         var count = mach_msg_type_number_t(MemoryLayout<mach_task_basic_info>.size) / 4
 
@@ -7752,73 +7752,18 @@ extension MessageCell {
 // The MessageInputViewDelegate methods have been moved to MessageInputHandler
 // for better separation of concerns.
 
-// Helper method for showing error alerts
-extension MessageableChannelViewController {
-    func showErrorAlert(message: String) {
-        let alert = UIAlertController(title: "Error", message: message, preferredStyle: .alert)
-        alert.addAction(UIAlertAction(title: "OK", style: .default))
-        present(alert, animated: true)
-    }
-}
+// MARK: - MessageableChannelViewController showErrorAlert Extension  (Moved to MessageableChannelViewController+Extensions.swift)
 
 // MARK: - UIImagePickerControllerDelegate Implementation
 // Note: UIImagePickerControllerDelegate is now handled by MessageInputHandler
 
-// MARK: - UITableViewDataSourcePrefetching
-extension MessageableChannelViewController: UITableViewDataSourcePrefetching {
-    func tableView(_ tableView: UITableView, prefetchRowsAt indexPaths: [IndexPath]) {
-        // Pre-cache message data for upcoming rows
-        for indexPath in indexPaths {
-            if indexPath.row < viewModel.messages.count {
-                let messageId = viewModel.messages[indexPath.row]
-                if let message = viewModel.viewState.messages[messageId],
-                    let author = viewModel.viewState.users[message.author]
-                {
+// MARK: - UITableViewDataSourcePrefetching tableView (Moved to MessageableChannelViewController+Extensions.swift)
 
-                    // Pre-load author's avatar
-                    let member = viewModel.getMember(message: message).wrappedValue
-                    let avatarInfo = viewModel.viewState.resolveAvatarUrl(
-                        user: author, member: member, masquerade: message.masquerade)
-
-                    // Fix: Only create the URL array if the URL is valid
-                    if let url = URL(string: avatarInfo.url.absoluteString) {
-                        // Use Kingfisher's ImagePrefetcher with the URL - make sure to not pass any arguments to start()
-                        let prefetcher = ImagePrefetcher(urls: [url])
-                        prefetcher.start()
-                    }
-
-                    // Pre-load message attachments if any
-                    if let attachments = message.attachments, !attachments.isEmpty {
-                        // Create an array to store valid attachment URLs
-                        let attachmentUrls = attachments.compactMap { attachment -> URL? in
-                            // Generate URL string and safely convert to URL object
-                            let urlString = viewModel.viewState.formatUrl(
-                                fromId: attachment.id, withTag: "attachments")
-                            return URL(string: urlString)
-                        }
-
-                        // Prefetch all attachments in one batch if there are any
-                        if !attachmentUrls.isEmpty {
-                            // Fix: Create the prefetcher and then start it - make sure to not pass any arguments to start()
-                            let prefetcher = ImagePrefetcher(urls: attachmentUrls)
-                            prefetcher.start()
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    func tableView(_ tableView: UITableView, cancelPrefetchingForRowsAt indexPaths: [IndexPath]) {
-        // Cancel pre-fetching for rows that are no longer needed
-        // Not critical to implement, but helps save resources
-    }
-}
-
+// MARK: - Empty State Handling showEmptyStateView()  (Moved to MessagableChannelViewController+Extensions.swift)
 // MARK: - Empty State Handling
 extension MessageableChannelViewController {
 
-    private func showEmptyStateView() {
+     private func showEmptyStateView() {
         // If the empty state view already exists, just make sure it's visible
         if let existingView = view.viewWithTag(100) {
             existingView.isHidden = false
@@ -7986,175 +7931,10 @@ extension MessageableChannelViewController {
     }
 }
 
-// MARK: - NSFWOverlayView and Delegate
-protocol NSFWOverlayViewDelegate: AnyObject {
-    func nsfwOverlayViewDidConfirm(_ view: NSFWOverlayView)
-}
 
-class NSFWOverlayView: UIView {
-    weak var delegate: NSFWOverlayViewDelegate?
+// MARK: - NSFWOverlayView and Delegate (Moved to NSFWOverlayView.swift)
 
-    private let channelName: String
-
-    // UI Components
-    private let overlayView = UIView()
-    private let stackView = UIStackView()
-    private let warningSymbol = UIImageView()
-    private let titleLabel = UILabel()
-    private let messageLabel = UILabel()
-    private let confirmButton = UIButton(type: .system)
-
-    init(channelName: String) {
-        self.channelName = channelName
-        super.init(frame: .zero)
-        setupUI()
-    }
-
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    // MARK: - UI Setup
-
-    private func setupUI() {
-        setupOverlay()
-        setupStackView()
-        setupWarningSymbol()
-        setupLabels()
-        setupConfirmButton()
-        setupConstraints()
-    }
-
-    private func setupOverlay() {
-        overlayView.backgroundColor = UIColor.black.withAlphaComponent(0.8)
-        overlayView.translatesAutoresizingMaskIntoConstraints = false
-        addSubview(overlayView)
-    }
-
-    private func setupStackView() {
-        stackView.axis = .vertical
-        stackView.alignment = .center
-        stackView.spacing = 16
-        stackView.translatesAutoresizingMaskIntoConstraints = false
-        overlayView.addSubview(stackView)
-    }
-
-    private func setupWarningSymbol() {
-        warningSymbol.image = UIImage(systemName: "exclamationmark.triangle.fill")
-        warningSymbol.contentMode = .scaleAspectFit
-        warningSymbol.tintColor = .textDefaultGray01
-        warningSymbol.translatesAutoresizingMaskIntoConstraints = false
-    }
-
-    private func setupLabels() {
-        // Title label
-        titleLabel.text = channelName
-        titleLabel.textColor = .textDefaultGray01
-        titleLabel.textAlignment = .center
-        titleLabel.font = UIFont.boldSystemFont(ofSize: 18)
-        titleLabel.numberOfLines = 0
-
-        // Message label
-        messageLabel.text = "This channel is marked as NSFW"
-        messageLabel.textColor = .textGray06
-        messageLabel.textAlignment = .center
-        messageLabel.font = UIFont.systemFont(ofSize: 14)
-        messageLabel.numberOfLines = 0
-    }
-
-    private func setupConfirmButton() {
-        confirmButton.setTitle("I confirm that I am at least 18 years old", for: .normal)
-        confirmButton.setTitleColor(.textDefaultGray01, for: .normal)
-        confirmButton.backgroundColor = UIColor.systemBlue
-        confirmButton.layer.cornerRadius = 8
-        confirmButton.contentEdgeInsets = UIEdgeInsets(top: 12, left: 16, bottom: 12, right: 16)
-        confirmButton.addTarget(self, action: #selector(confirmButtonTapped), for: .touchUpInside)
-        confirmButton.titleLabel?.numberOfLines = 0
-        confirmButton.titleLabel?.textAlignment = .center
-    }
-
-    private func setupConstraints() {
-        NSLayoutConstraint.activate([
-            // Overlay fills entire view
-            overlayView.leadingAnchor.constraint(equalTo: leadingAnchor),
-            overlayView.trailingAnchor.constraint(equalTo: trailingAnchor),
-            overlayView.topAnchor.constraint(equalTo: topAnchor),
-            overlayView.bottomAnchor.constraint(equalTo: bottomAnchor),
-
-            // Stack view centered in overlay
-            stackView.centerXAnchor.constraint(equalTo: overlayView.centerXAnchor),
-            stackView.centerYAnchor.constraint(equalTo: overlayView.centerYAnchor),
-            stackView.widthAnchor.constraint(equalTo: overlayView.widthAnchor, multiplier: 0.8),
-
-            // Warning symbol size
-            warningSymbol.heightAnchor.constraint(equalToConstant: 100),
-            warningSymbol.widthAnchor.constraint(equalToConstant: 100),
-        ])
-
-        // Add arranged subviews
-        stackView.addArrangedSubview(warningSymbol)
-        stackView.addArrangedSubview(titleLabel)
-        stackView.addArrangedSubview(messageLabel)
-        stackView.addArrangedSubview(confirmButton)
-    }
-
-    // MARK: - Actions
-
-    @objc private func confirmButtonTapped() {
-        delegate?.nsfwOverlayViewDidConfirm(self)
-    }
-
-    // MARK: - Public Methods
-
-    func show(in parentView: UIView, animated: Bool = true) {
-        translatesAutoresizingMaskIntoConstraints = false
-        parentView.addSubview(self)
-
-        NSLayoutConstraint.activate([
-            leadingAnchor.constraint(equalTo: parentView.leadingAnchor),
-            trailingAnchor.constraint(equalTo: parentView.trailingAnchor),
-            topAnchor.constraint(equalTo: parentView.topAnchor),
-            bottomAnchor.constraint(equalTo: parentView.bottomAnchor),
-        ])
-
-        if animated {
-            alpha = 0
-            UIView.animate(withDuration: 0.3) {
-                self.alpha = 1
-            }
-        }
-    }
-
-    func dismiss(animated: Bool = true, completion: (() -> Void)? = nil) {
-        if animated {
-            UIView.animate(
-                withDuration: 0.3,
-                animations: {
-                    self.alpha = 0
-                }
-            ) { _ in
-                self.removeFromSuperview()
-                completion?()
-            }
-        } else {
-            removeFromSuperview()
-            completion?()
-        }
-    }
-
-    // MARK: - Static Convenience Method
-
-    static func show(in parentView: UIView, channelName: String, delegate: NSFWOverlayViewDelegate?)
-        -> NSFWOverlayView
-    {
-        let overlay = NSFWOverlayView(channelName: channelName)
-        overlay.delegate = delegate
-        overlay.show(in: parentView, animated: true)
-        return overlay
-    }
-}
-
-// Add simple notification banner to show messages to the user
+// Add simple notification banner to show messages to the user 
 private class NotificationBanner {
     private let containerView = UIView()
     private let messageLabel = UILabel()
@@ -8221,478 +8001,13 @@ private class NotificationBanner {
     }
 }
 
+
+
 // Add these extensions at the end of the file, just before the last curly brace
 
-// Replace the existing extension for MessageCell with this one
-extension MessageCell {
-    private struct AssociatedKeys {
-        static var reactionsEnabledKey = "reactionsEnabled"
-    }
-
-    // Add this property to MessageCell to control reaction button visibility
-    var reactionsEnabled: Bool {
-        get {
-            // Get associated object or return default value
-            return objc_getAssociatedObject(self, &AssociatedKeys.reactionsEnabledKey) as? Bool
-                ?? true
-        }
-        set {
-            // Store new value using associated object
-            objc_setAssociatedObject(
-                self,
-                &AssociatedKeys.reactionsEnabledKey,
-                newValue,
-                .OBJC_ASSOCIATION_RETAIN_NONATOMIC
-            )
-            // Update UI based on the new value
-            updateReactionsVisibility()
-        }
-    }
-
-    // Update the visibility of reaction buttons
-    private func updateReactionsVisibility() {
-        // Update any reaction buttons based on the enabled state
-        reactionButton?.isHidden = !reactionsEnabled
-    }
-
-    // Add a property for the reaction button - you'll need to implement this
-    private var reactionButton: UIButton? {
-        // Return the actual reaction button from your view hierarchy
-        return subviews.compactMap { $0 as? UIButton }.first(where: {
-            $0.accessibilityIdentifier == "reactionButton"
-        })
-    }
-}
-
-// Add the uploadButtonEnabled property to MessageInputView class
-extension MessageInputView {
-    private struct AssociatedKeys {
-        static var uploadButtonEnabledKey = "uploadButtonEnabled"
-    }
-
-    // Add this property to control upload button visibility
-    var uploadButtonEnabled: Bool {
-        get {
-            // Return stored value or default
-            return objc_getAssociatedObject(self, &AssociatedKeys.uploadButtonEnabledKey) as? Bool
-                ?? true
-        }
-        set {
-            // Store the value
-            objc_setAssociatedObject(
-                self,
-                &AssociatedKeys.uploadButtonEnabledKey,
-                newValue,
-                .OBJC_ASSOCIATION_RETAIN_NONATOMIC
-            )
-            // Update the upload button visibility
-            uploadButton?.isHidden = !newValue
-        }
-    }
-
-    // Optional upload button reference - implement as needed
-    var uploadButton: UIButton? {
-        // Return the actual upload button from your view hierarchy
-        return subviews.compactMap { $0 as? UIButton }.first(where: {
-            $0.accessibilityIdentifier == "uploadButton"
-        })
-    }
-}
+// MARK: - Extension MessageCell (Moved to MessageableChannelViewController+MessageCell)
 
 // MARK: - UITextViewDelegate (moved to extension)
-
-// MARK: - MessageSkeletonView
-class MessageSkeletonView: UIView {
-
-    private let numberOfSkeletons = 9
-    private let skeletonRowHeight: CGFloat = 60
-
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        setupSkeletonView()
-    }
-
-    required init?(coder: NSCoder) {
-        super.init(coder: coder)
-        setupSkeletonView()
-    }
-
-    private func setupSkeletonView() {
-        backgroundColor = .bgDefaultPurple13
-
-        let stackView = UIStackView()
-        stackView.axis = .vertical
-        stackView.spacing = 12
-        stackView.translatesAutoresizingMaskIntoConstraints = false
-        addSubview(stackView)
-
-        NSLayoutConstraint.activate([
-            stackView.topAnchor.constraint(equalTo: topAnchor, constant: 20),
-            stackView.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 16),
-            stackView.trailingAnchor.constraint(equalTo: trailingAnchor, constant: -16),
-            stackView.bottomAnchor.constraint(lessThanOrEqualTo: bottomAnchor, constant: -20),
-        ])
-
-        // Create skeleton rows
-        for _ in 0..<numberOfSkeletons {
-            let skeletonRow = createSkeletonRow()
-            stackView.addArrangedSubview(skeletonRow)
-        }
-    }
-
-    private func createSkeletonRow() -> UIView {
-        let rowContainer = UIView()
-        rowContainer.translatesAutoresizingMaskIntoConstraints = false
-
-        // Avatar placeholder
-        let avatarView = UIView()
-        avatarView.backgroundColor = UIColor.systemGray5.withAlphaComponent(0.8)
-        avatarView.layer.cornerRadius = 20
-        avatarView.translatesAutoresizingMaskIntoConstraints = false
-        rowContainer.addSubview(avatarView)
-
-        // Content container
-        let contentStack = UIStackView()
-        contentStack.axis = .vertical
-        contentStack.spacing = 6
-        contentStack.translatesAutoresizingMaskIntoConstraints = false
-        rowContainer.addSubview(contentStack)
-
-        // Username placeholder
-        let usernameView = UIView()
-        usernameView.backgroundColor = UIColor.systemGray4.withAlphaComponent(0.7)
-        usernameView.layer.cornerRadius = 4
-        usernameView.translatesAutoresizingMaskIntoConstraints = false
-        contentStack.addArrangedSubview(usernameView)
-
-        // Message content placeholder
-        let messageView = UIView()
-        messageView.backgroundColor = UIColor.systemGray5.withAlphaComponent(0.6)
-        messageView.layer.cornerRadius = 4
-        messageView.translatesAutoresizingMaskIntoConstraints = false
-        contentStack.addArrangedSubview(messageView)
-
-        NSLayoutConstraint.activate([
-            rowContainer.heightAnchor.constraint(equalToConstant: skeletonRowHeight),
-
-            // Avatar constraints
-            avatarView.leadingAnchor.constraint(equalTo: rowContainer.leadingAnchor),
-            avatarView.topAnchor.constraint(equalTo: rowContainer.topAnchor, constant: 8),
-            avatarView.widthAnchor.constraint(equalToConstant: 40),
-            avatarView.heightAnchor.constraint(equalToConstant: 40),
-
-            // Content stack constraints
-            contentStack.leadingAnchor.constraint(equalTo: avatarView.trailingAnchor, constant: 12),
-            contentStack.trailingAnchor.constraint(equalTo: rowContainer.trailingAnchor),
-            contentStack.topAnchor.constraint(equalTo: rowContainer.topAnchor, constant: 8),
-
-            // Username placeholder constraints
-            usernameView.heightAnchor.constraint(equalToConstant: 16),
-            usernameView.widthAnchor.constraint(equalToConstant: 120),
-
-            // Message content placeholder constraints
-            messageView.heightAnchor.constraint(equalToConstant: 20),
-        ])
-
-        return rowContainer
-    }
-}
-
-// MARK: - Additional Memory Management
-extension MessageableChannelViewController {
-
-    override func viewDidDisappear(_ animated: Bool) {
-        super.viewDidDisappear(animated)
-
-        print(
-            "⚡ VIEW_DID_DISAPPEAR: User has completely left channel \(viewModel.channel.id) - performing FINAL instant cleanup"
-        )
-        let finalCleanupStartTime = CFAbsoluteTimeGetCurrent()
-
-        // Check if we're returning from search - if so, don't cleanup
-        if isReturningFromSearch {
-            print("🔍 VIEW_DID_DISAPPEAR: Returning from search, skipping final cleanup")
-            return
-        }
-
-        // CRITICAL FIX: Don't cleanup if we're navigating to the same channel with a target message
-        if let targetId = viewModel.viewState.currentTargetMessageId {
-            // Check if we're staying in the same channel
-            let isStayingInSameChannel: Bool
-            if case .channel(let currentChannelId) = viewModel.viewState.currentChannel {
-                isStayingInSameChannel = currentChannelId == viewModel.channel.id
-            } else {
-                isStayingInSameChannel = false
-            }
-
-            if isStayingInSameChannel {
-                print("🎯 VIEW_DID_DISAPPEAR: Staying in same channel, skipping final cleanup")
-                return
-            }
-        }
-
-        // CRITICAL FIX: Invalidate scroll check timer to prevent memory leak
-        scrollCheckTimer?.invalidate()
-        scrollCheckTimer = nil
-
-        // CRITICAL FIX: Cleanup MessageInputView references to prevent memory leaks
-        messageInputView?.cleanup()
-
-        // IMMEDIATE FINAL CLEANUP: No delays, no async operations
-        performFinalInstantCleanup()
-
-        // AGGRESSIVE: Force immediate memory cleanup when view disappears
-        forceImmediateMemoryCleanup()
-
-        // Clear preloaded status so it can be preloaded again when needed
-        let channelId = viewModel.channel.id
-        viewModel.viewState.preloadedChannels.remove(channelId)
-        print("🧹 CLEANUP: Cleared preloaded status for channel \(channelId)")
-
-        let finalCleanupEndTime = CFAbsoluteTimeGetCurrent()
-        let finalCleanupDuration = (finalCleanupEndTime - finalCleanupStartTime) * 1000
-        print(
-            "⚡ VIEW_DID_DISAPPEAR: Total final cleanup completed in \(String(format: "%.2f", finalCleanupDuration))ms"
-        )
-
-        // Log final memory usage
-        logMemoryUsage(prefix: "FINAL CLEANUP COMPLETE")
-    }
-
-    // Force immediate memory cleanup - called after view disappears
-    private func forceImmediateMemoryCleanup() {
-        print("⚡ FORCE_IMMEDIATE_CLEANUP: Starting INSTANT memory cleanup")
-        let startTime = CFAbsoluteTimeGetCurrent()
-
-        // IMMEDIATE: Force image cache cleanup
-        let cache = ImageCache.default
-        cache.clearMemoryCache()
-
-        // IMMEDIATE: Aggressive user cleanup - NO Task, NO async
-        let channelId = self.viewModel.channel.id
-        let isDM = self.viewModel.channel.isDM
-        let isGroupDM = self.viewModel.channel.isGroupDmChannel
-
-        print(
-            "👥 INSTANT_USER_CLEANUP: Starting for channel \(channelId) - DM: \(isDM), GroupDM: \(isGroupDM)"
-        )
-        let initialUserCount = self.viewModel.viewState.users.count
-
-        // Collect all user IDs that should be kept
-        var usersToKeep = Set<String>()
-
-        // Add current user if exists
-        if let currentUserId = self.viewModel.viewState.currentUser?.id {
-            usersToKeep.insert(currentUserId)
-        }
-
-        // Only keep users from OTHER channels (not the one we just left)
-        for (otherChannelId, messageIds) in self.viewModel.viewState.channelMessages {
-            // Skip the channel we just left
-            if otherChannelId == channelId { continue }
-
-            // Add users from messages in other channels
-            for messageId in messageIds {
-                if let message = self.viewModel.viewState.messages[messageId] {
-                    usersToKeep.insert(message.author)
-                    if let mentions = message.mentions {
-                        usersToKeep.formUnion(mentions)
-                    }
-                }
-            }
-        }
-
-        // IMMEDIATE: Keep users from OTHER active DMs
-        for channel in self.viewModel.viewState.channels.values {
-            // Skip the channel we just left
-            if channel.id == channelId { continue }
-
-            // Keep users from other active DMs
-            if channel.isDM || channel.isGroupDmChannel {
-                let recipientIds = channel.recipients
-                usersToKeep.formUnion(recipientIds)
-                print(
-                    "👥 INSTANT_USER_CLEANUP: Keeping \(recipientIds.count) users from other DM \(channel.id)"
-                )
-            }
-        }
-
-        // IMMEDIATE: Keep users that might be needed for server lists
-        for server in self.viewModel.viewState.servers.values {
-            usersToKeep.insert(server.owner)
-        }
-
-        print("👥 INSTANT_USER_CLEANUP: Users to keep: \(usersToKeep.count)")
-
-        // IMMEDIATE: For DMs, be more aggressive about user cleanup
-        if isDM || isGroupDM {
-            print("👥 INSTANT_USER_CLEANUP: Performing DM-specific user cleanup")
-
-            // Get users from the DM we just left
-            let dmRecipients = self.viewModel.channel.recipients
-            let usersToRemove = dmRecipients.filter { userId in
-                !usersToKeep.contains(userId)
-            }
-
-            if !usersToRemove.isEmpty {
-                print(
-                    "👥 INSTANT_USER_CLEANUP: Removing \(usersToRemove.count) DM users that are no longer needed"
-                )
-                for userId in usersToRemove {
-                    self.viewModel.viewState.users.removeValue(forKey: userId)
-                }
-            } else {
-                print("👥 INSTANT_USER_CLEANUP: All DM users are still needed elsewhere")
-            }
-        }
-
-        let finalUserCount = self.viewModel.viewState.users.count
-        let endTime = CFAbsoluteTimeGetCurrent()
-        let duration = (endTime - startTime) * 1000
-
-        print(
-            "⚡ FORCE_IMMEDIATE_CLEANUP: Completed in \(String(format: "%.2f", duration))ms - Users: \(initialUserCount) -> \(finalUserCount)"
-        )
-    }
-
-    /// Performs INSTANT final cleanup with no delays
-    private func performFinalInstantCleanup() {
-        let channelId = viewModel.channel.id
-        print("⚡ FINAL_INSTANT_CLEANUP: Starting IMMEDIATE final cleanup for channel \(channelId)")
-
-        let startTime = CFAbsoluteTimeGetCurrent()
-
-        // 1. IMMEDIATE: Clear table view synchronously
-        self.tableView.dataSource = nil
-        self.tableView.delegate = nil
-
-        // 2. IMMEDIATE: Force ViewState cleanup synchronously (no Task, no async)
-        viewModel.viewState.cleanupChannelFromMemory(
-            channelId: channelId, preserveForNavigation: false)
-        viewModel.viewState.forceMemoryCleanup()
-
-        // 3. IMMEDIATE: Aggressive image cache cleanup
-        ImageCache.default.clearMemoryCache()
-
-        // 4. IMMEDIATE: Reset all controller state
-        targetMessageId = nil
-        targetMessageProcessed = false
-        isInTargetMessagePosition = false
-        isLoadingMore = false
-        messageLoadingState = .notLoading
-
-        // 5. IMMEDIATE: Final verification and force cleanup
-        let remainingMessages = viewModel.viewState.messages.values.filter {
-            $0.channel == channelId
-        }.count
-        let remainingChannelMessages = viewModel.viewState.channelMessages[channelId]?.count ?? 0
-
-        if remainingMessages > 0 || remainingChannelMessages > 0 {
-            print(
-                "⚠️ FINAL_INSTANT_CLEANUP: Found \(remainingMessages) remaining messages, force removing"
-            )
-
-            // IMMEDIATE: Force remove any remaining data
-            viewModel.viewState.channelMessages.removeValue(forKey: channelId)
-
-            let finalMessagesToRemove = viewModel.viewState.messages.keys.filter { messageId in
-                if let message = viewModel.viewState.messages[messageId] {
-                    return message.channel == channelId
-                }
-                return false
-            }
-
-            for messageId in finalMessagesToRemove {
-                viewModel.viewState.messages.removeValue(forKey: messageId)
-            }
-
-            print(
-                "⚡ FINAL_INSTANT_CLEANUP: Force removed \(finalMessagesToRemove.count) remaining messages"
-            )
-        }
-
-        // 6. IMMEDIATE: Force garbage collection
-        _ =
-            viewModel.viewState.messages.count + viewModel.viewState.users.count
-            + viewModel.viewState.channelMessages.count
-
-        let endTime = CFAbsoluteTimeGetCurrent()
-        let duration = (endTime - startTime) * 1000  // Convert to milliseconds
-
-        print(
-            "⚡ FINAL_INSTANT_CLEANUP: Completed in \(String(format: "%.2f", duration))ms - ALL memory freed immediately!"
-        )
-        logMemoryUsage(prefix: "AFTER INSTANT FINAL CLEANUP")
-    }
-}
-
-// MARK: - Scroll Position Preservation
-extension MessageableChannelViewController {
-
-    /// Reloads the table view while maintaining the user's scroll position using message IDs as anchors
-    private func reloadTableViewMaintainingScrollPosition(messagesForDataSource: [String]) {
-        guard let visibleIndexPaths = tableView.indexPathsForVisibleRows,
-            !visibleIndexPaths.isEmpty
-        else {
-            // No visible rows, just reload normally
-            tableView.reloadData()
-            return
-        }
-
-        // Find an anchor message ID from visible rows (prefer middle visible row for stability)
-        var anchorMessageId: String?
-        var anchorDistanceFromTop: CGFloat = 0
-
-        // Try to find a good anchor from the middle of visible rows
-        let middleIndex = visibleIndexPaths.count / 2
-        for (index, indexPath) in visibleIndexPaths.enumerated() {
-            // Prefer rows that are not at the very edges
-            if index >= middleIndex && indexPath.row < messagesForDataSource.count {
-                anchorMessageId = messagesForDataSource[indexPath.row]
-                let cellFrame = tableView.rectForRow(at: indexPath)
-                anchorDistanceFromTop = cellFrame.origin.y - tableView.contentOffset.y
-                // print("🔍 SCROLL_PRESERVE: Selected anchor message \(anchorMessageId!) at index \(indexPath.row), distance from top: \(anchorDistanceFromTop)")
-                break
-            }
-        }
-
-        // Fallback to first visible row if no middle row found
-        if anchorMessageId == nil, let firstVisible = visibleIndexPaths.first,
-            firstVisible.row < messagesForDataSource.count
-        {
-            anchorMessageId = messagesForDataSource[firstVisible.row]
-            let cellFrame = tableView.rectForRow(at: firstVisible)
-            anchorDistanceFromTop = cellFrame.origin.y - tableView.contentOffset.y
-            // print("🔍 SCROLL_PRESERVE: Using fallback anchor message \(anchorMessageId!) at index \(firstVisible.row)")
-        }
-
-        // Perform the reload
-        tableView.reloadData()
-        tableView.layoutIfNeeded()
-
-        // Restore position to the anchor message
-        if let anchorId = anchorMessageId {
-            // Find the anchor message in the new data
-            if let newIndex = messagesForDataSource.firstIndex(of: anchorId) {
-                let newIndexPath = IndexPath(row: newIndex, section: 0)
-                let newCellFrame = tableView.rectForRow(at: newIndexPath)
-                let newContentOffsetY = newCellFrame.origin.y - anchorDistanceFromTop
-
-                // Ensure the offset is within valid bounds
-                let maxOffset = max(
-                    0,
-                    tableView.contentSize.height - tableView.bounds.height
-                        + tableView.contentInset.bottom)
-                let clampedOffset = max(0, min(newContentOffsetY, maxOffset))
-
-                tableView.setContentOffset(CGPoint(x: 0, y: clampedOffset), animated: false)
-                // print("📍 SCROLL_PRESERVE: Restored position to anchor message at new index \(newIndex), offset: \(clampedOffset)")
-            } else {
-                // print("⚠️ SCROLL_PRESERVE: Could not find anchor message \(anchorId) in new data")
-            }
-        }
-    }
-}
 
 // MARK: - Helper Functions
 /// Generates a dynamic message link based on the current domain
